Update Lift commands and constants so that encoder values increase when raising
- ensure encoder values are integers
- ensure WristDown and LowerLift default motor speed is in other direction
  as WristUp and RaiseLift default motor speed

+++ src/main/java/frc/robot/commands/WristUp.java
@@ -13,8 +13,8 @@ import frc.robot.util.Constants;
 
 public class WristUp extends Command {
   double motorPowerPercent;
-  double minEncoderPosThreshold; // Encoder values decrease when wrist goes up
-  double midEncoderPosThreshold;
+  int minEncoderPosThreshold; // Encoder values decrease when wrist goes up
+  int midEncoderPosThreshold;
 
   public WristUp() {
     // Use requires() here to declare subsystem dependencies
